backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
tensorflow==2.13.0